Unit tests still failing.  Problem is Realm.initializeSynchedRealm:385, check for requiresMigration
