+++ realm/realm-library/src/androidTest/java/io/realm/rule/TestRealmConfigurationFactory.java
@@ -16,10 +16,26 @@
 
 package io.realm.rule;
 
+import android.content.Context;
+import android.content.res.AssetManager;
+import android.support.test.InstrumentationRegistry;
+
+import org.junit.rules.TemporaryFolder;
+import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
+
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
+import io.realm.TestHelper;
 
 import static org.junit.Assert.assertTrue;
 
@@ -29,14 +45,62 @@ import static org.junit.Assert.assertTrue;
  * {@link Realm#deleteRealm(RealmConfiguration)} will throw an exception in the {@link #after()} method.
  * The temp directory will be deleted regardless if the {@link Realm#deleteRealm(RealmConfiguration)} fails or not.
  */
-public class TestRealmConfigurationFactory extends BaseConfigurationFactory {
+public class TestRealmConfigurationFactory extends TemporaryFolder {
+    private Map<RealmConfiguration, Boolean> map = new ConcurrentHashMap<RealmConfiguration, Boolean>();
+    private Set<RealmConfiguration> configurations = Collections.newSetFromMap(map);
+    protected boolean unitTestFailed = false;
+
+    @Override
+    public Statement apply(final Statement base, Description description) {
+        return new Statement() {
+            @Override
+            public void evaluate() throws Throwable {
+                before();
+                try {
+                    base.evaluate();
+                } catch (Throwable throwable) {
+                    unitTestFailed = true;
+                    throw throwable;
+                } finally {
+                    after();
+                }
+            }
+        };
+    }
+
+    @Override
+    protected void before() throws Throwable {
+        super.before();
+        Realm.init(InstrumentationRegistry.getTargetContext());
+    }
+
+    @Override
+    protected void after() {
+        // Wait all async tasks done to ensure successful deleteRealm call.
+        // This will throw when timeout. And the reason of timeout needs to be solved properly.
+        TestHelper.waitRealmThreadExecutorFinish();
+
+        try {
+            for (RealmConfiguration configuration : configurations) {
+                Realm.deleteRealm(configuration);
+            }
+        } catch (IllegalStateException e) {
+            // Only throw the exception caused by deleting the opened Realm if the test case itself doesn't throw.
+            if (!unitTestFailed) {
+                throw e;
+            }
+        } finally {
+            // This will delete the temp directory.
+            super.after();
+        }
+    }
 
     public RealmConfiguration createConfiguration() {
         RealmConfiguration configuration = new RealmConfiguration.Builder()
                 .directory(getRoot())
                 .build();
 
-        registerConfig(configuration);
+        configurations.add(configuration);
         return configuration;
     }
 
@@ -48,7 +112,7 @@ public class TestRealmConfigurationFactory extends BaseConfigurationFactory {
                 .name(name)
                 .build();
 
-        registerConfig(configuration);
+        configurations.add(configuration);
         return configuration;
     }
 
@@ -58,7 +122,7 @@ public class TestRealmConfigurationFactory extends BaseConfigurationFactory {
                 .name(name)
                 .build();
 
-        registerConfig(configuration);
+        configurations.add(configuration);
         return configuration;
     }
 
@@ -69,11 +133,48 @@ public class TestRealmConfigurationFactory extends BaseConfigurationFactory {
                 .encryptionKey(key)
                 .build();
 
-        registerConfig(configuration);
+        configurations.add(configuration);
         return configuration;
     }
 
     public RealmConfiguration.Builder createConfigurationBuilder() {
         return new RealmConfiguration.Builder().directory(getRoot());
+    }
+
+    // Copies a Realm file from assets to temp dir
+    public void copyRealmFromAssets(Context context, String realmPath, String newName) throws IOException {
+        RealmConfiguration config = new RealmConfiguration.Builder()
+                .directory(getRoot())
+                .name(newName)
+                .build();
+
+        copyRealmFromAssets(context, realmPath, config);
+    }
+
+    public void copyRealmFromAssets(Context context, String realmPath, RealmConfiguration config) throws IOException {
+        // Delete the existing file before copy
+        Realm.deleteRealm(config);
+
+        File outFile = new File(config.getRealmDirectory(), config.getRealmFileName());
+
+        InputStream is = null;
+        FileOutputStream os = null;
+        try {
+            is = context.getAssets().open(realmPath);
+            os = new FileOutputStream(outFile);
+
+            byte[] buf = new byte[1024];
+            int bytesRead;
+            while ((bytesRead = is.read(buf)) > -1) {
+                os.write(buf, 0, bytesRead);
+            }
+        } finally {
+            if (is != null) {
+                try { is.close(); } catch (IOException ignore) {}
+            }
+            if (os != null) {
+                try { os.close(); } catch (IOException ignore) {}
+            }
+        }
     }
 }
