--- realm/realm-library/src/androidTest/java/io/realm/rule/TestSyncConfigurationFactory.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2017 Realm Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.realm.rule;
-
-import io.realm.SyncConfiguration;
-import io.realm.SyncUser;
-
-public class TestSyncConfigurationFactory extends BaseConfigurationFactory {
-
-    public SyncConfiguration.Builder createConfigurationBuilder(SyncUser user, String uri) {
-        return new SyncConfiguration.Builder(user, uri).directory(getRoot());
-    }
-}
